requirements.txt
speedtest-cli
requests
pytz
python-dotenv==1.0.0
